// shared/src/interfaces/userInterface.ts
/**
 * @fileoverview Defines the IUser interface for user data.
 * @version 1.0.0
 * @module userInterface
 */

import { Optional } from 'sequelize';

/**
 * Interface representing the attributes of a user.
 */
export interface IUser {
  id?: number;
  avatarUrl?: object;
  email: string;
  mobileNumber: string;
  password: string;
  role?: 'admin' | 'candidate' | 'recruiter';
  isVerified?: boolean;
  createdAt?: Date;
  updatedAt?: Date;
}

/**
 * Interface for creating a new user (excluding autogenerated fields).
 */
export type IUserCreation = Optional<IUser, 'id' | 'createdAt' | 'updatedAt'>;